refactor: adjust layout
